Changes are in the description.
- [cloud] added Firebase Functions handlers for
 - Facebook token updates: updates the user profile if any of the
information has changed. Also used for "registering".
 - User info changes: We create a summary of users profile for other
users to consume (no last name or exact birthdate for example). Also
makes the app lighter.
 - Mobile numbers: When a mobile number is added to the users account,
we immediately start the verification process (send SMS etc.) (missing
validations)
 - Network (university) e-mails: Recently we've switched to Firebase
Auth verifications but they were a bit buggy and sometimes the
verification screen just hanged. Also there was no way to customize the
experince (read: the e-mail). This works similiar to the mobile number
verifications. When the e-mail is updated, we validate the e-mail,
network, if the network is allowed and make changes to the db
accordingly. We send an e-mail with a custom link which user can click.
The link verifies the user and app is updated in real-time. Also
reports malicious activity (read: user tries to update manually by
proxing our traffic.) The page user is redirected is in the /code/site
directory.
 - User photo changes: When an avatar or profile loop changes, we push
it to the public summary. NOTE: only works for avatars now since loop
uploading is WIP.
- [web] changed app in Firebase since I wanted to get rid of the
"hello-somerandomkey" project name. It's now "proco-app".
- [mobile] cleaner variables in the App container.