again, big changes, check the description.
So, since gave up on creating users with Facebook and their test users,
I’ve replaced the functions inside the parrot with new ones. We can now
create n number of dolls (test users) and spread them across a
location. I’ve tested with 1000 users and our pool seems to work just
all right. There is only a problem with the isSeenCurrentQuestion rule,
which is caused by user not having a question. You can’t be in the pool
without a question so it’s not a major problem.

I’ve also gave up on using React on the map UI (lost hours on that
shit), just created a simple HTML page with no CSS and some JS. Works
for testing. We should probably show the current user filters and such.
(Tip: If you wanna test it, you are gonna need a token, just go to
Parrot and run “acl generatetoken <uid>” (uid is who you wanna control)

Right now, 2 major things are missing: These test users don’t have a
loop or a question. That’s the next thing I’m adding and then we can
test the app.